Revised code to run CVA6

+++ fv/synthlc/step11_run_instn.sh
@@ -68,17 +68,20 @@ STEP 11 at $(pwd) $(date)
 # 2. ...
 DIR=xFollowerSetsOnly
 PYSCRPT=xFollowerSetsOnly
-if [ -d "${DIR}" ]; then 
-    echo "Directory exists ${DIR} and skip"
-#else 
+confirmed="y"
+if [ -d "${DIR}" ]; then
+    echo "Directory exists $INAME. Redo step? [y/n]"
+    read confirmed
+fi
+if [ $confirmed == "y" ]; then
     cp -r ../${DIR} .
 
    
     JOB1=rtl2mupath_followers
     JOB2=rtl2mupath_first_pls
     JOB3=rtl2mupath_first_pl_sets
-    TCLFILE=$(realpath ${DIR})/${JOB1}.tcl
-    SVFILE=$(realpath ${DIR})/${JOB1}.sv
+    TCLFILE1=$(realpath ${DIR})/${JOB1}.tcl
+    SVFILE1=$(realpath ${DIR})/${JOB1}.sv
     TCLFILE2=$(realpath ${DIR})/${JOB2}.tcl
     SVFILE2=$(realpath ${DIR})/${JOB2}.sv
     TCLFILE3=$(realpath ${DIR})/${JOB3}.tcl
@@ -88,19 +91,21 @@ if [ -d "${DIR}" ]; then
     python3 ${PYSCRPT}.py gen; 
 
     cd ../../..
-    ./run.sh ${FV_UNITDIR} ${TCLFILE} ${SVFILE}
-
+    #./run.sh ${FV_UNITDIR} ${TCLFILE1} ${SVFILE1}
+    ./RUN_JG.sh -j ${INAME_DIR}/${DIR} -s ${SVFILE1} -t ${TCLFILE1} -g 1
     cd ${INAME_DIR}/${DIR};
     python3 ${PYSCRPT}.py gen_s2; 
 
     cd ../../..
-    ./run.sh ${FV_UNITDIR} ${TCLFILE2} ${SVFILE2}
+    #./run.sh ${FV_UNITDIR} ${TCLFILE2} ${SVFILE2}
+    ./RUN_JG.sh -j ${INAME_DIR}/${DIR} -s ${SVFILE2} -t ${TCLFILE2} -g 1
 
     cd ${INAME_DIR}/${DIR};
     python3 ${PYSCRPT}.py gen_s3; 
 
     cd ../../..
-    ./run.sh ${FV_UNITDIR} ${TCLFILE3} ${SVFILE3}
+    #./run.sh ${FV_UNITDIR} ${TCLFILE3} ${SVFILE3}
+    ./RUN_JG.sh -j ${INAME_DIR}/${DIR} -s ${SVFILE3} -t ${TCLFILE3} -g 1
 
     cd ${INAME_DIR}/${DIR};
     python3 ${PYSCRPT}.py pp
